+++ src/components/sidebar/Categories.tsx
@@ -1,40 +1,15 @@
-"use client";
-
-import { useRecoilState } from "recoil";
-import { categoryAtom } from "@/libs/atoms";
-import Link from "next/link";
-import { usePathname } from "next/navigation";
+import CategoryTag from "./CategoryTag";
 
 type CategoriesProps = {
   categories: string[];
 };
 
 export default function Categories({ categories }: CategoriesProps) {
-  const [selectedCategory, setSelectedCategory] = useRecoilState(categoryAtom);
-  const pathname = usePathname();
-  // const test = pathname.startsWith("/React");
-  console.log(selectedCategory);
-
-  const bgColor = (category: string) => {
-    return selectedCategory === category ? "bg-teal-400" : "";
-  };
-
   return (
     <div className="text-center my-10">
-      <button
-        className={`border rounded-xl px-2 mr-1.5 my-2 ${bgColor("ALL")}`}
-        onClick={() => setSelectedCategory("ALL")}
-      >
-        ALL
-      </button>
+      <CategoryTag category="ALL" />
       {categories.sort().map((category, i) => (
-        <button
-          key={i}
-          className={`border rounded-xl px-2 mr-1.5 my-2 ${bgColor(category)}`}
-          onClick={() => setSelectedCategory(category)}
-        >
-          {category}
-        </button>
+        <CategoryTag key={i} category={category} />
       ))}
     </div>
   );
